chore(protocol-impl): move repository data objects to protocol-impl
