+++ broker-core/src/main/java/io/zeebe/broker/workflow/repository/ListWorkflowsControlMessageHandler.java
@@ -22,6 +22,9 @@ import io.zeebe.msgpack.value.ValueArray;
 import io.zeebe.protocol.Protocol;
 import io.zeebe.protocol.clientapi.ControlMessageType;
 import io.zeebe.protocol.clientapi.ErrorCode;
+import io.zeebe.protocol.impl.data.repository.ListWorkflowsControlRequest;
+import io.zeebe.protocol.impl.data.repository.ListWorkflowsResponse;
+import io.zeebe.protocol.impl.data.repository.WorkflowMetadata;
 import io.zeebe.protocol.impl.record.RecordMetadata;
 import io.zeebe.transport.ServerOutput;
 import io.zeebe.util.buffer.BufferUtil;
