+++ protocol-impl/src/main/java/io/zeebe/protocol/impl/data/repository/ListWorkflowsControlRequest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright © 2017 camunda services GmbH (info@camunda.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.zeebe.protocol.impl.data.repository;
+
+import io.zeebe.msgpack.UnpackedObject;
+import io.zeebe.msgpack.property.StringProperty;
+import org.agrona.DirectBuffer;
+
+public class ListWorkflowsControlRequest extends UnpackedObject {
+  private final StringProperty bpmnProcessIdProp = new StringProperty("bpmnProcessId", "");
+
+  public ListWorkflowsControlRequest() {
+    this.declareProperty(bpmnProcessIdProp);
+  }
+
+  public DirectBuffer getBpmnProcessId() {
+    return bpmnProcessIdProp.getValue();
+  }
+
+  public ListWorkflowsControlRequest setBpmnProcessId(final DirectBuffer directBuffer) {
+    bpmnProcessIdProp.setValue(directBuffer);
+    return this;
+  }
+}
